get rid of datastructures module
This leads to decomposition of functions in algorithms:
  order_packages() decomposited to:
  - group_packages_by_sources()
  - order_groupped_packages_by_arches()

order_groupped_packages_by_arches() needs indexed version of
ORDERED_ARCHES, so it have became dictionary. It is needed to speed up
sorted() in the algorithm (if it is a list, then the complexity will be
O(n) not O(1) with a dictionary).

For the sake of completeness added sort_groupped_packages_by_sources() if
you want to sort unordered groupped packages (notice that rdb provides
already sorted by sources packages).

The version is bumped to 3.0.0 due to API is broken by this commit.